Rewrite with Dijkstras algorithm
The problem is that it still sucks balls. It must be faster, but it
can't be *that* much faster.

My guess (needs profiling) is that trying to find the minimum node is
most of the running time. It's basically V^2 because I'm comparing based
on a map lookup, and the map has no index based on V. I can probably
speed it up by pulling all nodes out of the map at once, and comparing
directly